[misc] Fix deprecated use
(cherry picked from commit 719e1c4d23d6be4398b6a1fc42e3d0fedc967791)

+++ xwiki-platform-core/xwiki-platform-cache/xwiki-platform-cache-infinispan/src/main/java/org/xwiki/cache/infinispan/internal/InfinispanConfigurationLoader.java
@@ -117,7 +117,7 @@ public class InfinispanConfigurationLoader extends AbstractCacheConfigurationLoa
                     || configuration.eviction().maxEntries() != maxEntries) {
                     builder = builder(builder, null);
                     builder.eviction().strategy(EvictionStrategy.LRU);
-                    builder.eviction().maxEntries(maxEntries);
+                    builder.eviction().maxEntries((long) maxEntries);
                 }
             }
